+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.1.0
 - Configure Maven project
-- 2022.12.03 Created methods for NPC class, now npc have defence and damage.All bad code removed (except for Character, KostjaTest, Test classes)
+- 2022.12.03 Created class fight and combatMechanics have some mistakes in Fight.fightMethod() with ending of method .
+- Created methods for NPC class, now npc have defence and damage.All bad code removed (except for Character, KostjaTest, Test classes)
 - 2022.12.02 The first choice of the action1() method works. Calculating damage and defence for player is working. Created file JavaLearner.